+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
+- Fix `ImageCropper` `onChange` firing on uploaded images.
+
+## [2.97.0] - 2020-11-16
+
+Feature:
+- Add `CreditCardForm` component.
+
+## [2.96.0] - 2020-11-12
+
+Feature:
+- Update `RewardCard` component.
 
 - Feature: Add `variant` prop to `VerticalStepper`, with default value `andromeda`.
 - Feature: Add `DotIcon` component.